basic test cleanups
