+++ src/animations/mod.rs
@@ -4,7 +4,6 @@ pub mod colors;
 pub mod core;
 pub mod epsilon;
 pub mod platform;
-pub mod platform_tests;
 pub mod spring;
 pub mod transform;
 pub mod tween;
